// src/declarations.d.ts
interface FormRecordFieldBase extends FormRecordBase {
  // CSS classes
  class?: string
  // property name of the field
  name: string
  input: FormRecordFieldInput
  rules?: Partial<{
    required: boolean
    minLength: number
    maxLength: number
  }>
}

/**
 * Boolean fields
 */
interface FormRecordFieldBoolean extends FormRecordFieldBase {
  input: 'boolean'
  display?: 'checkbox' | 'toggle'
}

/**
 * Text fields
 */
interface FormRecordFieldOnelineText extends FormRecordFieldBase {
  input: 'oneline-text'
  type?: 'email' | 'url' | 'tel' | 'text'
}
interface FormRecordFieldMultilineText extends FormRecordFieldBase {
  input: 'multiline-text'
}

/**
 * Number fields
 */
interface FormRecordFieldNumber extends FormRecordFieldBase {
  input: 'number'
}
interface FormRecordFieldFloat extends FormRecordFieldBase {
  input: 'float'
}

/**
 * Date fields
 */
interface FormRecordFieldDate extends FormRecordFieldBase {
  input: 'date'
  display?: 'input' | 'calendar'
}
interface FormRecordFieldDatetime extends FormRecordFieldBase {
  input: 'datetime'
}

/**
 * Field with limited choices
 */
type SingleDataValue = FormRecordBase & {
  value: string
  default?: boolean
}
interface FormRecordFieldSingleData extends FormRecordFieldBase {
  input: 'single-data'
  display?: 'radio' | 'dropdown'
  values: SingleDataValue[]
}
interface FormRecordFieldMultiData extends FormRecordFieldSingleData {
  input: 'multi-data'
  display?: 'checkbox' | 'dropdown'
}

/**
 * Field with choices from external ressource
 */
interface FormRecordFieldSingleRelatedData extends FormRecordFieldBase {
  input: 'single-related-data'
  display: 'dropdown'
  endpoint: string
  valueField: string
  labelField: string
}
interface FormRecordFieldMultiRelatedData extends FormRecordFieldSingleRelatedData {
  input: 'multi-related-data'
}

type FormRecordField =
  FormRecordFieldBoolean |
  FormRecordFieldOnelineText | FormRecordFieldMultilineText |
  FormRecordFieldNumber | FormRecordFieldFloat |
  FormRecordFieldDate | FormRecordFieldDatetime |
  FormRecordFieldSingleData | FormRecordFieldMultiData |
  FormRecordFieldSingleRelatedData | FormRecordFieldMultiRelatedData